# backend/OPENAPI_SPEC.yaml
openapi: 3.0.3
info:
  title: Directory Wizard API
  description: |
    Multi-tenant Directory Creation Wizard API for creating and managing directory websites
    through a guided interface with integrated external service provisioning.
  version: 1.0.0
  contact:
    name: Directory Wizard Support
    email: support@directorywizard.com
    url: https://docs.directorywizard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.directorywizard.com/api
    description: Production server
  - url: https://staging-api.directorywizard.com/api
    description: Staging server
  - url: http://localhost:3001/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /wizard/{sessionId}:
    get:
      summary: Get wizard session
      description: Retrieve wizard session details
      tags:
        - Wizard
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Wizard session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WizardSessionResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /wizard/{sessionId}/step:
    put:
      summary: Update wizard step
      description: Update wizard step and session data
      tags:
        - Wizard
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WizardStepRequest'
      responses:
        '200':
          description: Wizard step updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WizardStepResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /wizard/{sessionId}/categories:
    post:
      summary: Process categories
      description: Process categories data in wizard context
      tags:
        - Wizard
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessCategoriesRequest'
      responses:
        '200':
          description: Categories processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessCategoriesResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /wizard/{sessionId}/listings:
    post:
      summary: Process listings
      description: Process listings data in wizard context
      tags:
        - Wizard
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessListingsRequest'
      responses:
        '200':
          description: Listings processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessListingsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /wizard/{sessionId}/complete:
    post:
      summary: Complete wizard
      description: Complete wizard and finalize tenant creation
      tags:
        - Wizard
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Wizard completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WizardCompleteResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        pattern: '^ses_[a-zA-Z0-9]+$'
      description: Wizard session ID
      example: ses_987654321
...
  schemas:
...
    WizardSessionResponse:
      type: object
      properties:
        id:
          type: string
          example: ses_987654321
        tenant_id:
          type: string
          example: tnt_123456789
        current_step:
          type: string
          enum: [BASIC_INFO, BRANDING, CATEGORIES, LISTINGS, PREVIEW, PUBLISH]
          example: BRANDING
        data:
          type: object
          description: Session data
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    WizardStepRequest:
      type: object
      required:
        - step
      properties:
        step:
          type: string
          enum: [BASIC_INFO, BRANDING, CATEGORIES, LISTINGS, PREVIEW, PUBLISH]
          example: BRANDING
        data:
          type: object
          description: Step-specific data
    WizardStepResponse:
      type: object
      properties:
        id:
          type: string
          example: ses_987654321
        current_step:
          type: string
          example: BRANDING
        data:
          type: object
          description: Updated session data
        updated_at:
          type: string
          format: date-time
    ProcessCategoriesRequest:
      type: object
      required:
        - categories
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryInput'
    CategoryInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Restaurants
        description:
          type: string
          example: Local dining establishments
        icon:
          type: string
          example: restaurant
        sort_order:
          type: integer
          example: 1
    ProcessCategoriesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        created_count:
          type: integer
          example: 2
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryOutput'
    CategoryOutput:
      type: object
      properties:
        id:
          type: string
          example: cat_123
        name:
          type: string
          example: Restaurants
        slug:
          type: string
          example: restaurants
        description:
          type: string
          example: Local dining establishments
        sort_order:
          type: integer
          example: 1
    ProcessListingsRequest:
      type: object
      required:
        - listings
      properties:
        listings:
          type: array
          items:
            $ref: '#/components/schemas/ListingInput'
    ListingInput:
      type: object
      required:
        - title
        - description
        - category
      properties:
        title:
          type: string
          example: Best Pizza Place
        description:
          type: string
          example: Authentic Italian pizza
        category:
          type: string
          example: Restaurants
        phone:
          type: string
          example: 555-0123
        email:
          type: string
          format: email
          example: info@pizza.com
        website:
          type: string
          format: uri
          example: https://pizza.com
    ProcessListingsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        created_count:
          type: integer
          example: 1
        listings:
          type: array
          items:
            $ref: '#/components/schemas/ListingOutput'
    ListingOutput:
      type: object
      properties:
        id:
          type: string
          example: lst_456
        title:
          type: string
          example: Best Pizza Place
        slug:
          type: string
          example: best-pizza-place
        description:
          type: string
          example: Authentic Italian pizza
        category:
          type: string
          example: Restaurants
        featured:
          type: boolean
          example: true
    WizardCompleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        tenant_id:
          type: string
          example: tnt_123456789
        message:
          type: string
          example: Wizard completed successfully
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        message:
          type: string
          example: Detailed error description
        field:
          type: string
          example: email
        code:
          type: string
          example: VALIDATION_ERROR
        timestamp:
          type: string
          format: date-time
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...